+++ src/algorithm/convex_hull_trick.rs
