+++ src/algorithm/fast_eratosthenes.rs
