+++ src/algorithm/inversion_number.rs
