+++ src/algorithm/largest_rectangle.rs
