+++ src/algorithm/mod.rs
@@ -1,12 +1,12 @@
 //! アルゴリズムっぽいものが置いてあります
 
 pub mod atkin;
-pub mod convexhulltrick;
+pub mod convex_hull_trick;
 pub mod cumsum;
-pub mod enumdivisors;
+pub mod enum_divisors;
 pub mod eratosthenes;
-pub mod fasteratosthenes;
-pub mod inversionnumber;
-pub mod largestrectangle;
+pub mod fast_eratosthenes;
+pub mod inversion_number;
+pub mod largest_rectangle;
 pub mod prime;
 pub mod run_length;
