+++ src/graph/mod.rs
@@ -1,5 +1,5 @@
 //! グラフに関係しそうなもの
 pub mod dijkstra;
-pub mod dijkstrarestorepath;
-pub mod floydwarshall;
+pub mod dijkstra_restorepath;
+pub mod floyd_warshall;
 pub mod util;
