+++ src/math/mod.rs
@@ -1,4 +1,4 @@
 //! 数学っぽいもの
 pub mod euclid;
-pub mod modpow;
+pub mod mod_pow;
 pub mod permutation;
