+++ src/string/mod.rs
@@ -1,3 +1,3 @@
 //! 文字列に関係しそうなもの
 pub mod rollinghash;
-pub mod zalgorithm;
+pub mod z_algorithm;
