+++ src/structure/binary_trie.rs
