+++ src/structure/cumsum_2d.rs
