+++ src/structure/fenwick_tree.rs
