+++ src/structure/segment_tree.rs
