+++ src/structure/sparse_table.rs
